stored activities in array
